access control posts

+++ app/Http/Controllers/PostsController.php
@@ -12,6 +12,17 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -76,6 +87,7 @@ class PostsController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
+
         return view('posts.show')->with('post', $post);
     }
 
@@ -88,6 +100,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for corect user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -123,6 +140,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for corect user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed');
     }
